+++ release-notes.md
@@ -1,6 +1,28 @@
 quicktill — cash register software
 ==================================
 
+Upgrade v20.x to v21
+--------------------
+
+What's new:
+
+ * Ability to specify minimum and/or maximum ABV for a department
+
+To upgrade the database:
+
+ - run psql and give the following commands to the database:
+
+```
+BEGIN;
+
+ALTER TABLE departments
+	ADD COLUMN minabv numeric(3,1),
+	ADD COLUMN maxabv numeric(3,1);
+
+COMMIT;
+```
+
+
 Upgrade v19.x to v20
 --------------------